+++ honey/lib/src/runner/default_functions.dart
@@ -1,37 +1,58 @@
-import 'package:honey/src/consts/function.dart';
-import 'package:honey/src/honey_widgets_binding.dart';
-import 'package:honey/src/runner/functions/action_functions.dart';
-import 'package:honey/src/runner/functions/core_functions.dart';
-import 'package:honey/src/runner/functions/date_functions.dart';
-import 'package:honey/src/runner/functions/logic_functions.dart';
-import 'package:honey/src/runner/functions/math_functions.dart';
-import 'package:honey/src/runner/functions/widget_functions.dart';
+import 'package:honey/honey.dart';
+import 'package:honey/src/runner/functions/and.dart';
+import 'package:honey/src/runner/functions/click.dart';
+import 'package:honey/src/runner/functions/contains.dart';
+import 'package:honey/src/runner/functions/divide.dart';
+import 'package:honey/src/runner/functions/ends_with.dart';
+import 'package:honey/src/runner/functions/enter.dart';
+import 'package:honey/src/runner/functions/equal.dart';
+import 'package:honey/src/runner/functions/error.dart';
+import 'package:honey/src/runner/functions/format.dart';
+import 'package:honey/src/runner/functions/greater.dart';
+import 'package:honey/src/runner/functions/less.dart';
+import 'package:honey/src/runner/functions/matches.dart';
+import 'package:honey/src/runner/functions/minus.dart';
+import 'package:honey/src/runner/functions/multiply.dart';
+import 'package:honey/src/runner/functions/not.dart';
+import 'package:honey/src/runner/functions/now.dart';
+import 'package:honey/src/runner/functions/or.dart';
+import 'package:honey/src/runner/functions/output.dart';
+import 'package:honey/src/runner/functions/plus.dart';
+import 'package:honey/src/runner/functions/pow.dart';
+import 'package:honey/src/runner/functions/property.dart';
+import 'package:honey/src/runner/functions/starts_with.dart';
+import 'package:honey/src/runner/functions/swipe.dart';
+import 'package:honey/src/runner/functions/variable.dart';
+import 'package:honey/src/runner/functions/verify.dart';
+import 'package:honey/src/runner/functions/wait.dart';
+import 'package:honey/src/runner/functions/widgets.dart';
 
-const defaultFunctions = <F, HoneyFunction>{
-  F.click: ActionFunctions.click,
-  F.verify: ActionFunctions.verify,
-  F.enter: ActionFunctions.enter,
-  F.wait: ActionFunctions.wait,
-  F.swipe: ActionFunctions.swipe,
-  F.output: ActionFunctions.output,
-  // F.error,
-  F.and: LogicFunctions.and,
-  F.or: LogicFunctions.or,
-  F.not: LogicFunctions.not,
-  F.widgets: WidgetFunctions.widgets,
-  F.property: CoreFunctions.property,
-  F.variable: CoreFunctions.variable,
-  //F.function,
-  F.now: DateFunctions.now,
-  F.format: DateFunctions.format,
-  F.equal: MathFunctions.equal,
-  F.greater: MathFunctions.greater,
-  F.less: MathFunctions.less,
-  F.plus: MathFunctions.plus,
-  F.minus: MathFunctions.minus,
-  F.multiply: MathFunctions.multiply,
-  F.divide: MathFunctions.divide,
-  F.pow: MathFunctions.pow,
-  //F.matches:Str.cat,
-  F.concat: CoreFunctions.concat,
+final defaultFunctions = <String, HoneyFunction>{
+  F.and.name: and,
+  F.click.name: click,
+  F.contains.name: contains,
+  F.divide.name: divide,
+  F.endsWith.name: endsWith,
+  F.enter.name: enter,
+  F.equal.name: equal,
+  F.error.name: error,
+  F.format.name: format,
+  F.greater.name: greater,
+  F.less.name: less,
+  F.matches.name: matches,
+  F.minus.name: minus,
+  F.multiply.name: multiply,
+  F.not.name: not,
+  F.now.name: now,
+  F.or.name: or,
+  F.output.name: output,
+  F.plus.name: plus,
+  F.pow.name: pow,
+  F.property.name: property,
+  F.startsWith.name: startsWith,
+  F.swipe.name: swipe,
+  F.variable.name: variable,
+  F.verify.name: verify,
+  F.wait.name: wait,
+  F.widgets.name: widgets,
 };